Moved Google Image Search logic out of view
