+++ QuoteDroplet/Models/GoogleImageModel.swift
@@ -1,0 +1,10 @@
+//
+//  GoogleImageModel.swift
+//  Quote Droplet
+//
+//  Created by Daniel Agapov on 2024-10-23.
+//
+
+import Foundation
+
+struct TestModel: Decodable {}
