+++ QuoteDroplet/ViewModels/AuthorViewModel.swift
@@ -24,6 +24,28 @@ class AuthorViewModel: ObservableObject {
         self.apiService = apiService
     }
     
+    func loadRemoteJSON<T: Decodable>(_ urlString: String, completion: @escaping  ((T) -> Void)) {
+        guard let url = URL(string: urlString) else {
+            fatalError("Invalid URL")
+        }
+        
+        let request = URLRequest(url: url)
+        URLSession.shared.dataTask(with: request) { data, response, error in
+            guard let data = data else {
+                fatalError(error?.localizedDescription ?? "Unknown Error")
+            }
+            
+            do {
+                let decoder = JSONDecoder()
+                let data = try decoder.decode(T.self, from: data)
+                print("data printed from loadremoteJSON")
+                print(data)
+                completion(data)
+            } catch {
+                fatalError("Couldn't parse data from \(urlString)\n\(error)")
+            }
+        }
+    }
     
     public func loadInitialQuotes() {
         totalQuotesLoaded = 0
